error handling and tunnnel

+++ controllers/mentorController.ts
@@ -24,27 +24,30 @@ const getMentor = (
       if (arr.length == 0) {
         return res.status(200).json({ mentor: null });
       }
-      User.findOne({ email: arr[0].mentoringUser[0].email }).then((user) => {
-        let mentorTrans = arr.map((mentor: IMentor[]) => {
-          return {
-            ...mentor,
-            mentoringUser: user,
-          };
+      User.findOne({ email: arr[0].mentoringUser[0].email })
+        .then((user) => {
+          if (!user) {
+            throw new http500Error("No user found");
+          }
+          let mentorTrans = arr.map((mentor: IMentor[]) => {
+            return {
+              ...mentor,
+              mentoringUser: user,
+            };
+          });
+          res.status(200).json({
+            mentor: {
+              ...mentorTrans[0]._doc,
+              mentoringUser: [mentorTrans[0].mentoringUser],
+            },
+          });
+        })
+        .catch((err) => {
+          next(err);
         });
-
-        res.status(200).json({
-          mentor: {
-            ...mentorTrans[0]._doc,
-            mentoringUser: [mentorTrans[0].mentoringUser],
-          },
-        });
-      });
     })
-    .catch((err: any) => {
-      if (!err.statusCode) {
-        err.statusCode = 500;
-      }
-      next(err);
+    .catch((error) => {
+      next(error);
     });
 };
 
@@ -57,12 +60,11 @@ const createMentor = (
   if (!errors.isEmpty()) {
     throw new http422Error(errors.array()[0].msg);
   }
-
   User.findOne({ email: req.body.user.email })
     .then((user: any) => {
       User.findOne({ email: req.body.email }).then((userMentor: any) => {
         const mentor = new Mentor({
-          name: req.body.name.trim(),
+          name: req.body.name,
           email: req.body.email,
           accepted: false,
           mentorId: userMentor?._id,
@@ -92,17 +94,17 @@ const createMentor = (
                   res.status(201).json({ mentor });
                 });
               })
-              .catch(() => {
-                next(new http500Error());
+              .catch((err) => {
+                next(err);
               });
           })
-          .catch(() => {
-            next(new http500Error());
+          .catch((err) => {
+            next(err);
           });
       });
     })
-    .catch(() => {
-      next(new http500Error());
+    .catch((err) => {
+      next(err);
     });
 };
 
@@ -111,33 +113,41 @@ const updateMentor = (
   res: Response<{ mentor: IMentor }>,
   next: NextFunction
 ) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    throw new http500Error(errors.array()[0].msg);
+  }
   Mentor.findByIdAndUpdate(req.params.id, req.body, { new: true })
     .then((mentor: any) => {
-      User.findOne({ _id: mentor.mentoringUserId }).then((user) => {
-        if (!user?.notificationToken) {
-          return res.status(201).json({ mentor });
-        }
-        expoNotification
-          .sendPushNotification({
-            to: user.notificationToken,
-            title: "Mentor",
-            body: `Mentor (${mentor.email}) accepted your request ✅`,
-          })
-          .then(() => {
-            res.status(201).json({ mentor });
-          })
-          .catch((err) => {
-            if (!err.statusCode) {
-              err.statusCode = 500;
-            }
-            next(err);
-          });
-      });
+      User.findOne({ _id: mentor.mentoringUserId })
+        .then((user) => {
+          if (!user) {
+            throw new http500Error("User doesn't exist");
+          }
+          if (!user?.notificationToken) {
+            return res.status(201).json({ mentor });
+          }
+          expoNotification
+            .sendPushNotification({
+              to: user.notificationToken,
+              title: "Mentor",
+              body: `Mentor (${mentor.email}) accepted your request ✅`,
+            })
+            .then(() => {
+              res.status(201).json({ mentor });
+            })
+            .catch((err) => {
+              if (!err.statusCode) {
+                err.statusCode = 500;
+              }
+              next(err);
+            });
+        })
+        .catch((err) => {
+          next(err);
+        });
     })
-    .catch((err: any) => {
-      if (!err.statusCode) {
-        err.statusCode = 500;
-      }
+    .catch((err) => {
       next(err);
     });
 };
@@ -147,20 +157,16 @@ const deleteMentor = (
   res: Response<{ success: any }>,
   next: NextFunction
 ) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    throw new http500Error(errors.array()[0].msg);
+  }
   Mentor.findOneAndDelete({ _id: req.params.id })
     .then((mentor: any) => {
-      return mentor;
+      Task.deleteMany({ mentorId: mentor.mentorId });
+      res.status(204);
     })
-    .then((mentor) => {
-      return Task.deleteMany({ mentorId: mentor.mentorId });
-    })
-    .then((result) => {
-      res.status(200).json({ success: "ok" });
-    })
-    .catch((err: any) => {
-      if (!err.statusCode) {
-        err.statusCode = 500;
-      }
+    .catch((err) => {
       next(err);
     });
 };
